Extract HappyMapper::ClassMethods to its own file
